Added feature to reject users while signup who will have to score marks in MDCAT above a certain threshold by calculating their required aggregate from their Matric and FSC Marks. Allowed users to log in even when the payment is pending to get the login token if they change the browser or device etc. Updated Payment Verification System to now enable users to update their plan from current plan also by just selecting their new plan and uploading the screenshot at page after /signup form, and enjoy their current plan until the payment is verified by the admins. Fixed a small bug in /quizzes/generate route.Implemented /users/me route, and updated getDashboardStats() function to match the current UI changes. Updated schema too to accomodate current API changes.

+++ backend/controllers/payment.js
@@ -4,7 +4,7 @@ import pool from "../database.js";
 import { unlink } from "node:fs/promises";
 
 export const getUnverifiedUsers = handleAsyncError(async (req, res, next) => {
-    const data = (await pool.query(" SELECT name, age, email, role, academic_status FROM users INNER JOIN students ON students.student_id=users.user_id WHERE payment_status='PENDING' ORDER BY user_id ASC")).rows;
+    const data = (await pool.query(" SELECT name, email, role, upgrade_role, academic_status, payment_status, upgrade_status, coupon FROM users INNER JOIN students ON students.student_id=users.user_id WHERE payment_status='PENDING' OR upgrade_status='PENDING' ORDER BY user_id ASC")).rows;
     res.status(200).json({
         status: "success",
         data
@@ -24,8 +24,13 @@ export const rejectPayment = handleAsyncError(async (req, res, next) => {
 export const verifyPayment = handleAsyncError(async (req, res, next) => {
     if (!req.body?.email)
         return next(new AppError("Please provide the user's email to verify his/her payment", 400));
-    await pool.query("UPDATE students SET payment_status='VERIFIED' WHERE student_id = (SELECT user_id FROM users WHERE email=$1)", [req.body.email]);
+    if (!req.body?.role)
+        return next(new AppError("Please provide user's new role to verify his/her payment", 400));
+
+    await pool.query("UPDATE students SET payment_status='VERIFIED', upgrade_status=NULL WHERE student_id = (SELECT user_id FROM users WHERE email=$1)", [req.body.email]);
+    await pool.query("UPDATE users SET role=$2 WHERE email=$1", [req.body.email, req.body.role]);
     await unlink(`receipts/${req.body.email}.jpg`);
+
     res.status(200).json({
         status: "success",
         message: "Payment Verified"
@@ -38,6 +43,14 @@ export const getPendingPaymentReceipt = (req, res, next) => {
 export const getPaymentStatus = (req, res, next) => {    
     res.status(200).json({
         status: "success",
-        payment_status: req.user.payment_status
+        payment_status: req.user.payment_status,
+        upgrade_status: req.user.upgrade_status
     });
 };
+
+export const verifyCoupon = handleAsyncError(async (req, res, next) => {
+    const result = (await pool.query("SELECT code FROM coupons WHERE code=$1", [req.body.coupon])).rows[0];
+    res.status(result ? 200 : 404).json({
+        status: result ? "success" : "fail"
+    });
+});
